+++ src/Event/Events/MutationAnalysis/MutationEvaluation/HeuristicSuppression/HeuristicSuppressionWasStarted.php
@@ -33,14 +33,17 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation;
+namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation\HeuristicSuppression;
 
-use Infection\Event\Subscriber\EventSubscriber;
+use Infection\Mutation\Mutation;
 
 /**
  * @internal
  */
-interface MutantEvaluationWasStartedSubscriber extends EventSubscriber
+final readonly class HeuristicSuppressionWasStarted
 {
-    public function onMutantEvaluationWasStarted(MutantEvaluationWasStarted $event): void;
+    public function __construct(
+        public Mutation $mutation,
+    ) {
+    }
 }
