+++ src/Event/Events/MutationAnalysis/MutationEvaluation/MutantAnalysis/MutantAnalysisWasStartedSubscriber.php
@@ -33,14 +33,14 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation;
+namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation\MutantAnalysis;
 
 use Infection\Event\Subscriber\EventSubscriber;
 
 /**
  * @internal
  */
-interface HeuristicSuppressionWasStartedSubscriber extends EventSubscriber
+interface MutantAnalysisWasStartedSubscriber extends EventSubscriber
 {
-    public function onHeuristicSuppressionWasStarted(HeuristicSuppressionWasStarted $event): void;
+    public function onMutantAnalysisWasStarted(MutantAnalysisWasStarted $event): void;
 }
