+++ src/Event/Events/MutationAnalysis/MutationEvaluation/MutantAnalysis/MutantEvaluation/MutantEvaluationWasFinishedSubscriber.php
@@ -33,14 +33,14 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation;
+namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation\MutantAnalysis\MutantEvaluation;
 
 use Infection\Event\Subscriber\EventSubscriber;
 
 /**
  * @internal
  */
-interface MutantMaterialisationWasFinishedSubscriber extends EventSubscriber
+interface MutantEvaluationWasFinishedSubscriber extends EventSubscriber
 {
-    public function onMutantMaterialisationWasFinished(MutantMaterialisationWasFinished $event): void;
+    public function onMutantEvaluationWasFinished(MutantEvaluationWasFinished $event): void;
 }
