+++ src/Event/Events/MutationAnalysis/MutationEvaluation/MutantAnalysis/MutantMaterialisation/MutantMaterialisationWasStarted.php
@@ -33,14 +33,17 @@
 
 declare(strict_types=1);
 
-namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation;
+namespace Infection\Event\Events\MutationAnalysis\MutationEvaluation\MutantAnalysis\MutantMaterialisation;
 
-use Infection\Event\Subscriber\EventSubscriber;
+use Infection\Mutant\Mutant;
 
 /**
  * @internal
  */
-interface MutantMaterialisationWasStartedSubscriber extends EventSubscriber
+final readonly class MutantMaterialisationWasStarted
 {
-    public function onMutantMaterialisationWasStarted(MutantMaterialisationWasStarted $event): void;
+    public function __construct(
+        public Mutant $mutant,
+    ) {
+    }
 }
